update5 in task-12

+++ 0x04-more_functions_nested_loops/101-print_number.c
@@ -5,9 +5,9 @@
  */
 void print_number(int n)
 {
-	long m;
+	int m;
 	int c;
-	long num;
+	int num;
 
 	num = n;
 	/* if negative */
